pyproject.toml
[project]
name = "weather-etl-pipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "EdissonMC",email = "pattrick.adamss@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pyarrow (>=19.0.1,<20.0.0)"
]

[tool.poetry]
packages = [{include = "weather_etl_pipeline", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
